refactor(look&feel,apollo): refactor ItemFile components for improved code clarity
