+++ client/apollo/react/src/Form/FileUpload/ItemFile/ItemFileLF.tsx
@@ -4,39 +4,20 @@ import { ClickIcon } from "../../../ClickIcon/ClickIconLF";
 import { Icon } from "../../../Icon/IconLF";
 import { Spinner } from "../../../Spinner/SpinnerLF";
 import { ItemMessage } from "../../ItemMessage/ItemMessageLF";
-import { ItemFileCommon } from "./ItemFileCommon";
-import type { ItemFileProps } from "./types";
+import { ItemFileCommon, type ItemFileProps } from "./ItemFileCommon";
 
 import "@axa-fr/design-system-apollo-css/dist/Form/FileUpload/ItemFile/ItemFileLF.css";
 
 export { itemFileVariants } from "./ItemFileCommon";
 
-export const ItemFile = ({
-  onDeleteClick,
-  onVisibilityClick,
-  ariaLabelVisibility,
-  ariaLabelDelete,
-  filename,
-  ...props
-}: ItemFileProps) => (
+export const ItemFile = (props: ItemFileProps) => (
   <ItemFileCommon
     ItemMessageComponent={ItemMessage}
     ItemIconComponent={Icon}
     ItemSpinnerComponent={Spinner}
-    deleteIcon={
-      <ClickIcon
-        src={deleteIconFilled}
-        onClick={onDeleteClick}
-        aria-label={ariaLabelDelete || `Suppression du fichier ${filename}`}
-      />
-    }
-    visibilityIcon={
-      <ClickIcon
-        src={visibilityIcon}
-        onClick={onVisibilityClick}
-        aria-label={ariaLabelVisibility || `Observer le fichier ${filename}`}
-      />
-    }
+    ClickIconComponent={ClickIcon}
+    deleteIcon={deleteIconFilled}
+    visibilityIcon={visibilityIcon}
     {...props}
   />
 );
