--- client/apollo/react/src/Form/FileUpload/ItemFile/types.ts
@@ -1,15 +1,0 @@
-import { type MouseEvent } from "react";
-import type { ItemFileCommonProps } from "./ItemFileCommon";
-
-export type ItemFileProps = Omit<
-  ItemFileCommonProps,
-  | "ItemMessageComponent"
-  | "ItemIconComponent"
-  | "ItemClickIconComponent"
-  | "ItemSpinnerComponent"
-  | "deleteIcon"
-  | "visibilityIcon"
-> & {
-  onDeleteClick?: (event: MouseEvent<HTMLButtonElement>) => void;
-  onVisibilityClick?: (event: MouseEvent<HTMLButtonElement>) => void;
-};
